requirements.txt
Flask==3.0.1
twilio==8.12.0
python-dotenv==1.0.1